Splash + Onboarding: refactor to better support splash screen + states
There's android's splash screen and my own splash screen which could be swapped out for an onboarding page.
Android 12 has some issues with launching splashscreen when hitting the run button so remember to launch it from the
app icon. Also newer versions of the navigation compose lib is slower when toggling startDestination. It tends to
flash the default route first.

Instead of having an if statement around the welcome screen to double verify, just delay isLoading to mask the logic
and slowness. Will need to revisit this implementation if we want to upgrade the nav lib