Welcome Screen: Finish animations
Messed around with a variety of animations. There's not a true sense of View.Invisible for compose so it's hard
to ensure that the layout is bounded and constrained correctly so view components don't shift
as they're added. Part of it could be that I'm laying out the objects incorrectly in the first place to prevent
shifting.

For now, added an AnimatedInvisibility which a play off of the existing ColumnScope.AnimatedVisibility
but just keeps the same composable with a 0f alpha underneath it to maintain its bounds.

tbd if using a counter to track the order of animations is ideal